// client/packages/editor-oss/src/controls/input/ActionTypes.ts
export type PlayerActions =
  | "jump"
  | "jump_hold"
  | "primary"
  | "primary_hold"
  | "secondary"
  | "secondary_hold"
  | "forward"
  | "lateral"
  | "run"
  | "crouch"
  | "reload"
  | "drop"
  | "use"
  | "pull"
  | "first_person_toggle"
  | "cam_lateral"
  | "cam_forward"
  | "cam_up_toggle"
  | "steer"
  | "view_x"
  | "view_y";